pyproject.toml
[tool.poetry]
name = "zero-test"
version = "0.1.0"
description = ""
authors = ["Federico <f.fulgeri@gmail.com>"]
readme = "README.md"
packages = [{include = "zero_test"}]

[tool.poetry.dependencies]
python = "^3.9"
pyzmq = "^24.0.0"
ipython = "^8.5.0"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
